fix build types in CMakeLists.txt
The command I used before to set the CMake build type, and the strings property for it, were not correct. The changes will properly set the build type if not set by the user, properly set the strings property, and respect a value set by the user via the command line.